requirements.txt
streamlit
python-dotenv
google-cloud-aiplatform
google-auth
google-generativeai
google-ai-generativelanguage
Pillow